third binomial

+++ Binomial.java
@@ -19,8 +19,8 @@ return binomial1(n - 1, k) + binomial1(n - 1, k - 1);
 	 }
 	
 	// Computes the Binomial function, efficiently
-	public static int binomial(int n, int k) {
-		int[][] memo= new int [n+1][k+1];
+	public static long binomial(int n, int k) {
+		long[][] memo= new long [n+1][k+1];
 		for(int i=0; i<=n;i++ ){
 		for(int j=0;j<=k;j++){
 				memo[i][j]=-1;
@@ -30,7 +30,7 @@ return binomial1(n - 1, k) + binomial1(n - 1, k - 1);
 		return binomial(n, k, memo);
 	}
 
-	private static int binomial(int n, int k, int[][] memo) {
+	private static long binomial(int n, int k, long[][] memo) {
 		if (memo[n][k] != -1) {
 			return memo[n][k];
 		}
